5. Various header-related build fixes for MacOSX 10.5 (Jaguar).  It is at       time unknown whether or not these negatively impact the build on older       versions of Darwin  (Marc La France).

+++ lib/GL/apple/dri_driver.c
@@ -1,4 +1,4 @@
-/* $XFree86: xc/lib/GL/apple/dri_driver.c,v 1.5tsi Exp $ */
+/* $XFree86: xc/lib/GL/apple/dri_driver.c,v 1.6tsi Exp $ */
 /**************************************************************************
 
 Copyright 1998-1999 Precision Insight, Inc., Cedar Park, Texas.
@@ -738,7 +738,7 @@ driCreatePixelFormat(Display *dpy, __DRIscreenPrivate *psp,
     int i;
     CGLPixelFormatAttribute attr[64]; // currently uses max of 30
     CGLPixelFormatObj result;
-    long n_formats;
+    GLint n_formats;
 
     i = 0;
